pyproject.toml
[project]
name = "sboxyourmom"
version = "1.0.0"
description = "An open-source SBox analysis utility"
requires-python = ">=3.7"
license = {file = "LICENSE"}

dependencies = [
    "pillow",
    "numpy"
]

[project.scripts]
sboxyourmom = "main:main"
